security controller separated
