+++ src/main/java/prodapp/User.java
@@ -6,20 +6,11 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.ManyToMany;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.ui.Model;
 
 @Entity
-@Configuration
-@EnableWebSecurity
-@EnableGlobalMethodSecurity( securedEnabled = true )
 public class User extends WebSecurityConfigurerAdapter{
 	@Id
 	@GeneratedValue
@@ -28,15 +19,6 @@ public class User extends WebSecurityConfigurerAdapter{
 
 	@ManyToMany(mappedBy = "users")
 	private Collection<Mission> missions;
-
-	@Autowired
-	protected void configureAuth(AuthenticationManagerBuilder auth) 
-			throws Exception {
-				auth.inMemoryAuthentication().withUser("admin").password("{noop}admin")
-				.roles("USER", "ADMIN")
-				.and().withUser("user").password("{noop}user").roles("USER");
-			}
-
 	
 	public User() {
 
@@ -63,14 +45,6 @@ public class User extends WebSecurityConfigurerAdapter{
 	public User(String userName, String contact) {
 		this.userName = userName;
 		this.contact = contact;
-	}
-	private String getLoggedInUser(Model model) {
-		Object principal = SecurityContextHolder.getContext()
-				.getAuthentication().getPrincipal();
-		if (principal instanceof User)
-			return ((User) principal).getUserName();
-		return principal.toString();
-		
 	}
 
 	@Override
